[mvsUtils] MultiViewParams: Move findTileNearestCams function from SgmDepthList
