+++ src/aliceVision/mvsUtils/MultiViewParams.hpp
@@ -10,6 +10,7 @@
 #include <aliceVision/mvsData/Point2d.hpp>
 #include <aliceVision/mvsData/Point3d.hpp>
 #include <aliceVision/mvsData/Pixel.hpp>
+#include <aliceVision/mvsData/ROI.hpp>
 #include <aliceVision/mvsData/StaticVector.hpp>
 #include <aliceVision/mvsData/structures.hpp>
 
@@ -294,6 +295,15 @@ public:
      */
     StaticVector<int> findNearestCamsFromLandmarks(int rc, int nbNearestCams) const;
 
+    /**
+     * @brief Find nearest cameras for a given tile
+     * @param[in] rc R camera id
+     * @param[in] nbNearestCams maximum number of desired nearest cameras
+     * @param[in] tCams a given list of pre-selected nearest cameras
+     * @param[in] roi the tile 2d region of interest
+     * @return nearest cameras list for the given tile
+     */
+    std::vector<int> findTileNearestCams(int rc, int nbNearestCams, const std::vector<int>& tCams, const ROI& roi) const;
 
     inline void setMinViewAngle(float minViewAngle)
     {
